Created the AssetLoader class with a Load() method and dispose() method. The Load() method contains 3 main objexts - Texture (Filter, Images, Resolution, Resizing), TextureRegion (Texture objects and their rectangular boundaries - bird, grass, background, skull, bar), Animation (Pass TextureRegion objects into constructor of Animation object; give the animation 3 frames).
Note that the images had to be flipped because I was using Y-down coordinate system, but libGDX uses Y-up coordinate system.

Added SpriteBatch to the GameRender class which helped with drawing the images. used the GameRender class to draw blocks to form grass, dirt, etc)

Added a runTime float to the render() methods, to determine which frame the bird animation should display

When I run the program, screen launches with the zombie bird. When you click the bird jumps up.

Had some issues with texture.png root directory. Hoepfully this image will also upload to Git.