Amended the Bird class to add functionality for the bird to rotate when rising and falling.
Created a Scrollable super class with 2 Grass objects and 3 Pipe objects. Used a ScrollHandler class as well in conjunction with these 2 classes, to set how the pipes will reset as the image scrolls left

Note: Creation of new objects in the game will add more to the heap, and the subroutine garbage collector which clears up this space may cause the game screen to lag or freeze. So I must remember not to create too many new objects.